final steps in merging

+++ src/com/apmath/loans/domain/models/ApplicationDetailsInterface.kt
@@ -5,6 +5,6 @@ import com.apmath.loans.domain.data.Money
 
 interface ApplicationDetailsInterface {
     val status: Status
-    val interest: Int
+    val interest: Float
     val term: Int
 }
