+++ src/com/apmath/loans/infrastructure/models/ApplicationDetails.kt
@@ -6,6 +6,6 @@ import com.apmath.loans.domain.data.Money
 
 data class ApplicationDetails (
     override val status: Status,
-    override val interest: Int,
+    override val interest: Float,
     override val term: Int
 ) : ApplicationDetailsInterface
